enhance: Refactor changelog page with modular components (SearchBar, ChangeBadge, PackageToggle)
