+++ includes/Admin/Dashboard/Dashboard.php
@@ -159,10 +159,13 @@ class Dashboard implements Hookable {
 
         // Get the dashboard page slug based on the legacy dashboard page settings.
         $is_legacy_dashboard_page = get_transient( 'dokan_legacy_dashboard_page' );
+        $is_legacy_changelog_page = get_transient( 'dokan_legacy_changelog_page' );
         $dashboard_page_slug      = $is_legacy_dashboard_page ? 'dokan' : 'dokan-dashboard';
+        $changelog_page_slug      = $is_legacy_changelog_page ? 'dokan' : 'dokan-dashboard';
         $dashboard_url            = admin_url( 'admin.php?page=' . $dashboard_page_slug );
+        $changelog_url            = admin_url( 'admin.php?page=' . $changelog_page_slug );
 
-        $header_info   = [
+        $header_info = [
             'lite_version'    => DOKAN_PLUGIN_VERSION,
             'is_pro_exists'   => dokan()->is_pro_exists(),
             'dashboard_url'   => $dashboard_url,
@@ -172,8 +175,8 @@ class Dashboard implements Hookable {
                 [
                     [
                         'id'       => 'whats-new',
-                        'title'    => esc_html__( "Changelog", 'dokan-lite' ),
-                        'url'      => $dashboard_url . '#/changelog',
+                        'title'    => esc_html__( 'Changelog', 'dokan-lite' ),
+                        'url'      => $changelog_url . '#/changelog',
                         'icon'     => 'whats-new',
                         'active'   => Helper::dokan_has_new_version(),
                         'external' => false,
