+++ src/admin/dashboard/pages/changelog/index.tsx
@@ -1,255 +1,9 @@
-import { __, sprintf } from '@wordpress/i18n';
-import { useState, useEffect, useCallback, useRef } from '@wordpress/element';
+import SearchBar from './SearchBar';
+import { __ } from '@wordpress/i18n';
+import VersionRow from './VersionRow';
 import apiFetch from '@wordpress/api-fetch';
-import { Search, X, Crown,Hammer,TrendingUp, Rocket,CircleCheck} from 'lucide-react';
-import { twMerge } from 'tailwind-merge';
-
-// Types
-interface ChangeItem {
-    title: string;
-    description: string;
-}
-
-interface VersionChanges {
-    [ key: string ]: ChangeItem[];
-}
-
-interface Version {
-    version: string;
-    released: string;
-    changes: VersionChanges;
-}
-
-// Badge component for change types
-const ChangeBadge = ( { type }: { type: string } ) => {
-    const getBadgeStyles = () => {
-        switch ( type ) {
-            case 'New':
-            case 'New Module':
-            case 'New Feature':
-                return 'bg-teal-500 text-white';
-            case 'Fix':
-                return 'bg-red-500 text-white';
-            case 'Improvement':
-            case 'Improvements':
-            case 'Update':
-                return 'bg-purple-500 text-white';
-            default:
-                return 'bg-blue-500 text-white';
-        }
-    };
-const renderIcon = () => {
-        // New Features / Modules
-        if ( [ 'New', 'New Module', 'New Feature' ].includes( type ) ) {
-            return <Rocket size={ 12 } />;
-        }
-        // Bug Fixes
-        if ( type === 'Fix' ) {
-            return <Hammer size={ 12 } />;
-        }
-        if ( [ 'Improvement', 'Improvements', 'Update' ].includes( type ) ) {
-            return <TrendingUp size={ 12 } />;
-        }
-        return null;
-    };
-    return (
-        <span
-            className={ `inline-flex items-center px-2 py-1 gap-[6px] rounded-[20px] text-xs font-semibold ${ getBadgeStyles() }` }
-        >
-            { renderIcon() }
-            { type }
-        </span>
-    );
-};
-
-// Search Component with Dropdown
-const SearchBar = ( {
-    versions,
-    onVersionSelect,
-}: {
-    versions: Version[];
-    onVersionSelect: ( index: number ) => void;
-} ) => {
-    const [ isOpen, setIsOpen ] = useState( false );
-    const [ searchQuery, setSearchQuery ] = useState( '' );
-    const searchRef = useRef< HTMLDivElement >( null );
-
-    useEffect( () => {
-        const handleClickOutside = ( event: MouseEvent ) => {
-            if (
-                searchRef.current &&
-                ! searchRef.current.contains( event.target as Node )
-            ) {
-                setIsOpen( false );
-            }
-        };
-
-        document.addEventListener( 'mousedown', handleClickOutside );
-        return () =>
-            document.removeEventListener( 'mousedown', handleClickOutside );
-    }, [] );
-
-    // Filter versions based on search query or show latest 5
-    const getDisplayVersions = () => {
-        if ( searchQuery.length > 0 ) {
-            return versions.filter( ( version ) =>
-                version.version
-                    .toLowerCase()
-                    .includes( searchQuery.toLowerCase() )
-            );
-        }
-        // Show latest 5 versions when no search query
-        return versions.slice( 0, 5 );
-    };
-
-    const displayVersions = getDisplayVersions();
-
-    return (
-        <div className="relative" ref={ searchRef }>
-            { /* Jump to version trigger button */ }
-            <button
-                onClick={ () => setIsOpen( ! isOpen ) }
-                className="flex items-center gap-1 text-sm text-[#575757] hover:text-gray-900"
-            >
-                { __( 'Jump to version', 'dokan-lite' ) }
-                <svg
-                    className={ `w-4 h-4 transition-transform ${
-                        isOpen ? 'rotate-180' : ''
-                    }` }
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                >
-                    <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={ 2 }
-                        d="M19 9l-7 7-7-7"
-                    />
-                </svg>
-            </button>
-
-            { /* Dropdown */ }
-            { isOpen && (
-                <div className="absolute top-full left-0 mt-2 w-64 bg-white rounded-lg shadow-lg border border-gray-200 z-50">
-                    { /* Search Input */ }
-                    <div className="px-4 py-2 border-b border-gray-200">
-                        <div className="relative">
-                            <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 w-4 h-4 text-gray-400" />
-                            <input
-                                type="text"
-                                value={ searchQuery }
-                                onChange={ ( e ) =>
-                                    setSearchQuery( e.target.value )
-                                }
-                                placeholder={ __(
-                                    'Search version…',
-                                    'dokan-lite'
-                                ) }
-                                className="w-full pl-10 pr-10 py-2 rounded-[3px] border border-[#7047EB] bg-white text-sm focus:outline-none focus:border-[#7047EB]"
-                            />
-                            { searchQuery && (
-                                <button
-                                    onClick={ () => setSearchQuery( '' ) }
-                                    className="absolute right-3 top-1/2 transform -translate-y-1/2 text-gray-400 hover:text-gray-600"
-                                >
-                                    <X className="w-4 h-4" />
-                                </button>
-                            ) }
-                        </div>
-                    </div>
-
-                    { /* Version List */ }
-                    <div className="max-h-64 overflow-y-auto py-1">
-                        { displayVersions.length > 0 ? (
-                            displayVersions.map( ( version ) => {
-                                const originalIndex = versions.findIndex(
-                                    ( v ) => v.version === version.version
-                                );
-                                const isMatch =
-                                    searchQuery.length > 0 &&
-                                    version.version
-                                        .toLowerCase()
-                                        .includes( searchQuery.toLowerCase() );
-                                return (
-                                    <button
-                                        key={ version.version }
-                                        onClick={ () => {
-                                            onVersionSelect( originalIndex );
-                                            setIsOpen( false );
-                                            setSearchQuery( '' );
-                                        } }
-                                        className={ `w-full text-left px-4 py-2 text-sm hover:bg-[#EFEAFF] focus:outline-none ${
-                                            isMatch
-                                                ? 'text-[#828282] hover:text-[#7047EB]'
-                                                : 'text-[#828282] hover:text-[#7047EB]'
-                                        }` }
-                                    >
-                                        { version.version }
-                                        { originalIndex === 0 && (
-                                            <span className="ml-2 text-xs text-gray-500">
-                                                (
-                                                { __( 'Latest', 'dokan-lite' ) }
-                                                )
-                                            </span>
-                                        ) }
-                                    </button>
-                                );
-                            } )
-                        ) : (
-                            <div className="px-4 py-2 text-sm text-gray-500">
-                                { __( 'No versions found', 'dokan-lite' ) }
-                            </div>
-                        ) }
-                    </div>
-                </div>
-            ) }
-        </div>
-    );
-};
-
-// Package Toggle (Lite/Pro)
-const PackageToggle = ( {
-    activePackage,
-    hasPro,
-    onToggle,
-}: {
-    activePackage: 'lite' | 'pro';
-    hasPro: boolean;
-    onToggle: ( pkg: 'lite' | 'pro' ) => void;
-} ) => {
-    if ( ! hasPro ) {
-        return null;
-    }
-
-    return (
-        <div className="inline-flex">
-            <button
-                onClick={ () => onToggle( 'lite' ) }
-                className={ `flex flex-col justify-center items-end px-[22px] py-2 text-sm font-semibold transition-colors rounded-l-[5px] border ${
-                    activePackage === 'lite'
-                        ? 'bg-[#7047EB] border-[#7047EB] text-white'
-                        : 'bg-white border-[#E9E9E9] text-black'
-                }` }
-            >
-                { __( 'Lite', 'dokan-lite' ) }
-            </button>
-            <button
-                onClick={ () => onToggle( 'pro' ) }
-                className={ `flex flex-col justify-center items-end px-[22px] py-2 text-sm font-medium transition-colors rounded-r-[5px] border ${
-                    activePackage === 'pro'
-                        ? 'bg-[#7047EB] border-[#7047EB] text-white'
-                        : 'bg-white border-[#E9E9E9] text-black'
-                }` }
-            >
-                <span className="flex items-center font-semibold gap-2">
-                    {__('Pro', 'dokan-lite')}
-                    <Crown size={ 16 } />
-                </span>
-            </button>
-        </div>
-    );
-};
+import PackageToggle from './PackageToggle';
+import { useState, useEffect, useCallback, useRef } from '@wordpress/element';
 
 // Loading Spinner
 const LoadingSpinner = () => (
@@ -257,150 +11,6 @@ const LoadingSpinner = () => (
         <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-blue-600"></div>
     </div>
 );
-
-// Version Row Component with sticky behavior
-const VersionRow = ( {
-    version,
-    isLatest,
-    isExpanded,
-    isHighlighted,
-    onToggle,
-}: {
-    version: Version;
-    isLatest: boolean;
-    isExpanded: boolean;
-    isHighlighted: boolean;
-    onToggle: () => void;
-} ) => {
-    const formatDate = ( dateString: string ) => {
-        const date = new Date( dateString );
-        return date.toLocaleDateString( 'en-US', {
-            day: 'numeric',
-            month: 'long',
-            year: 'numeric',
-        } );
-    };
-
-    const changeTypes = Object.keys( version.changes );
-
-    // Calculate total items across all change types
-    const totalItems = changeTypes.reduce(
-        ( sum, type ) => sum + version.changes[ type ].length,
-        0
-    );
-
-    // Show toggle only when total items > 5 (not for latest)
-    const shouldShowToggle = ! isLatest && totalItems > 5;
-
-    // For non-latest versions, show limited content when collapsed
-    const getVisibleChanges = () => {
-        if ( isLatest || isExpanded || totalItems <= 5 ) {
-            return version.changes;
-        }
-
-        // Show only first 5 items total when collapsed
-        const limitedChanges: VersionChanges = {};
-        let itemCount = 0;
-        const maxItems = 5;
-
-        for ( const type of changeTypes ) {
-            if ( itemCount >= maxItems ) {
-                break;
-            }
-
-            const items = version.changes[ type ];
-            const remainingSlots = maxItems - itemCount;
-            const itemsToShow = items.slice( 0, remainingSlots );
-
-            if ( itemsToShow.length > 0 ) {
-                limitedChanges[ type ] = itemsToShow;
-                itemCount += itemsToShow.length;
-            }
-        }
-
-        return limitedChanges;
-    };
-
-    const visibleChanges = getVisibleChanges();
-    const visibleChangeEntries = Object.entries( visibleChanges );
-
-    return (
-        <div className="flex gap-20 mb-10">
-            { /* Left Column - Version Info (Sticky) */ }
-            <div className="w-48 flex-shrink-0">
-                <div className="sticky top-4">
-                    <div className="flex items-center gap-1.5 mb-1">
-                        <h3 className="text-lg font-semibold text-gray-900">
-                            { version.version }
-                        </h3>
-                        { isLatest && (
-                            <span className="inline-flex items-center">
-                                <span className='block border-l border-[#A5A5AA] h-4'></span>
-                                <span className='flex items-center px-2 py-1 rounded-full text-sm font-normal text-[#008864] gap-1.5'>
-                                    { __( 'Latest', 'dokan-lite' ) }
-                                    <CircleCheck size={12} />
-                                </span>
-                            </span>
-                        ) }
-                    </div>
-                    <p className="text-sm text-gray-500">
-                        { formatDate( version.released ) }
-                    </p>
-                </div>
-            </div>
-
-            { /* Right Column - Changes Card */ }
-            <div className="flex-1">
-                <div
-                    className={ `bg-white rounded-lg transition-all duration-300 shadow-[0_1px_2px_-1px_rgba(0,0,0,0.10),0_1px_3px_0_rgba(0,0,0,0.10)] ${
-                        isHighlighted
-                            ? 'ring-2 ring-blue-500 ring-offset-2'
-                            : 'border border-gray-200'
-                    }` }
-                >
-                    { visibleChangeEntries.map(
-                        ( [ type, items ], typeIndex ) => (
-                            <div key={ type }>
-                                { /* Add border/divider before each section except the first */ }
-                                { typeIndex > 0 && (
-                                    <hr className="my-6 w-full h-px bg-[#E9E9E9] border-0 m-0" />
-                                ) }
-                                <div className='p-6'>
-                                    <ChangeBadge type={ type } />
-                                    <ul className="mt-3 space-y-3">
-                                        { items.map( ( item, itemIndex ) => (
-                                            <li
-                                                key={ itemIndex }
-                                                className="flex items-start text-sm text-[#575757] mb-0"
-                                            >
-                                                <span className="mr-2 text-sm text-[#575757]">
-                                                    •
-                                                </span>
-                                                <span>{ item.title }</span>
-                                            </li>
-                                        ) ) }
-                                    </ul>
-                                </div>
-                            </div>
-                        )
-                    ) }
-
-                    { /* See Details / See Less Toggle */ }
-                    { shouldShowToggle && (
-                        <button
-                            onClick={ onToggle }
-                            className="mt-4 text-sm text-[#7047EB] hover:text-[#5a38bc] font-medium !outline-none !focus:outline-none focus:ring-0 focus:border-0 underline"
-                        >
-                            { isExpanded
-                                ? __( 'See Less', 'dokan-lite' )
-                                : __( 'See Details', 'dokan-lite' ) }
-                        </button>
-                    ) }
-                </div>
-            </div>
-        </div>
-    );
-};
 
 // Main Changelog Page Component
 const ChangelogPage = () => {
